API fini , rectification d'erreur s'il en a

+++ controllers/stuff.js
@@ -1,4 +1,5 @@
 const Thing = require("../modeles/Thing");
+const fs = require("fs");
 
 exports.createThing = (req, res, next) => {
     //on instance le modele et passe en parametre le corps de la requêtes
@@ -51,9 +52,20 @@ exports.updateThing = (req, res, next) => {
 };
 
 exports.deleteOneThing = (req, res, next) => {
-    Thing.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: "Objet supprimé !" }))
-        .catch((error) => res.status(400).json({ error }));
+    Thing.findOne({ _id: req.params.userId }).then((thing) => {
+        if (thing.userId != req.auth.userId) {
+            res.status(401).json({ message: "Non autorisé!" });
+        } else {
+            const filename = thing.imageUrl.split("images/")[1];
+            fs.unlink(`images/${filename}`, () => {
+                Thing.deleteOne({ _id: req.params.id })
+                    .then(() =>
+                        res.status(200).json({ message: "Objet suprimé" })
+                    )
+                    .catch((error) => res.status(500).json({ error }));
+            });
+        }
+    });
 };
 
 exports.getOneThing = (req, res, next) => {
